Added tests.

+++ tests/generate.py
@@ -2,6 +2,9 @@ import shutil
 import subprocess
 import itertools
 import os
+import json
+
+import numpy as np
 
 
 light_command = """
@@ -23,15 +26,31 @@ def generatePOVFile(filename, template, light_position):
         ))
 
 
-def generateImages(template, light_positions):
+def generateImages(template, light_positions, lightning_file="lightning.json"):
+    """Generates images using POV-Ray and lightning file"""
+    devnull = open(os.devnull, 'w')
+    base_name = template.split(".")[0]
     filenames = []
+    lightning = {}
     for light_position in light_positions:
-        filename = "{0}-{1}.pov".format(template.split(".")[0], ".".join(map(str, light_position)))
+        if np.linalg.norm(light_position) < 10:
+            print("Light seems too close of the object.")
+        filename = "{0}-{1}.pov".format(base_name, ".".join(map(str, light_position)))
         generatePOVFile(filename, template, light_position)
-        subprocess.call(["povray", filename])
+        output = subprocess.call(["povray", filename], stderr=devnull)
+        if output:
+            raise Exception('Could not execute povray.')
         os.remove(filename)
-        filenames.append(filename.rsplit(".", maxsplit=1)[0] + ".png")
-    return filenames
+        filenames.append(filename.rsplit(".", 1)[0] + ".png")
+
+        lightning[filenames[-1]] = np.negative(np.array(light_position, dtype=np.float64))
+        lightning[filenames[-1]] /= np.linalg.norm(lightning[filenames[-1]])
+        lightning[filenames[-1]] = lightning[filenames[-1]].tolist()
+
+    with open(lightning_file, 'w') as fhdl:
+        json.dump(lightning, fhdl)
+
+    return filenames, lightning_file
 
 
 if __name__ == '__main__':
@@ -39,6 +58,6 @@ if __name__ == '__main__':
     posy = [-20, 0, 20]
     posz = [-20, 0, 20]
     light_positions = itertools.product(posx, posy, posz)
-    files = generateImages("sphere.pov.tmpl", light_positions)
+    files = generateImages("cube_angled.pov.tmpl", light_positions)
     for file_ in files:
         os.remove(file_)
